server/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==1.3.0
Django==3.2.18
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
frozenlist==1.4.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
multidict==6.0.4
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pkgutil_resolve_name==1.3.10
promise==2.3
psycopg2-binary==2.9.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.2
zipp==3.15.0
